Added unit tests
